Update and rename local_variable.cs to local_variables.cs
